backend/requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
python-decouple==3.8
pandas         
numpy
redis
django-filter
celery>=5.3.0
django-filter
drf-spectacular